Add `IMMEDIATE` policy for tests that used to rely on default EOT prior we changed the behavior in killbill/killbill#1739
With the new behavior, `EOT` policy is now honored and so if we want to keep the current assertions, we have to overwrite such policy with `IMMEDIATE` (to match prev behavior).

An alternate fix would be to keep `EOT` (default policy) and fix the tests to adjust to new behavior.